requirements.txt
Python 3.7+
Streamlit
OpenCV
NumPy
Pillow
DeepFace
